# [core] CPD: Optimize --skip-lexical-errors option