--- a/pmd-core/src/main/java/net/sourceforge/pmd/cpd/TokenEntry.java
+++ b/pmd-core/src/main/java/net/sourceforge/pmd/cpd/TokenEntry.java
@@ -4,8 +4,8 @@
 
 package net.sourceforge.pmd.cpd;
 
-import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -95,20 +95,22 @@ public class TokenEntry implements Comparable<TokenEntry> {
      */
     public static class State {
         private int tokenCount;
-        private Map<String, Integer> tokens;
-        private List<TokenEntry> entries;
+        private int tokensMapSize;
 
-        public State(List<TokenEntry> entries) {
+        public State() {
             this.tokenCount = TokenEntry.TOKEN_COUNT.get().intValue();
-            this.tokens = new HashMap<>(TokenEntry.TOKENS.get());
-            this.entries = new ArrayList<>(entries);
+            this.tokensMapSize = TokenEntry.TOKENS.get().size();
         }
 
-        public List<TokenEntry> restore() {
+        public void restore(List<TokenEntry> entries) {
             TokenEntry.TOKEN_COUNT.get().set(tokenCount);
-            TOKENS.get().clear();
-            TOKENS.get().putAll(tokens);
-            return entries;
+            Iterator<Map.Entry<String, Integer>> it = TOKENS.get().entrySet().iterator();
+            while (it.hasNext()) {
+                if (it.next().getValue() > tokensMapSize) {
+                    it.remove();
+                }
+            }
+            entries.subList(tokenCount, entries.size()).clear();
         }
     }