@jsotuyod 2021-07-06
This looks great. I'd probably review the semantics, but certainly a valid approach and shown, the reduced allocations should greatly improve the performance.